backend/pyproject.toml
[tool.poetry]
name = "api_guide"
version = "0.1.0"
description = ""
authors = ["4got10king <Aldavletshina77654@yandex.ru>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.9"
alembic = "^1.14.1"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
aiosqlite = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"